% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaProjectHG38.generic.R
\docType{class}
\name{TrenaProjectHG38.generic-class}
\alias{TrenaProjectHG38.generic-class}
\alias{.TrenaProjectHG38.generic}
\alias{TrenaProjectHG38.generic}
\title{TrenaProjectHG38.generic-class}
\usage{
TrenaProjectHG38.generic(quiet = TRUE)
}
\value{
An object of the TrenaProjectHG38.generic class
}
\description{
Expression, variant and covariate data for the genes of interest (perhaps unbounded) for pre-term birth studies
}
